First commit
This is the first commit. I made a simple program that consists of a main() method and a chooseWord() method. The choose word method outputs a random word from a text file containing around 200 words. So far the only thing the main method does is assign the output of the chooseWord() method to a String variable called 'word', and prints it to the terminal.