2nd commit to the hangman project
This is the 2nd commit to my hangman project, I added a new method called startGame() that takes the random word as an argument. Once you call this method, it creates a functional hangman game(even though the only hangman aspect is messages that tell you how many body parts he has) with exception handling(This is done with nested if/else statements that print appropriate "error messages" to the console. Next session I plan to actually throw proper exceptions with messages). The only major thing left to do with this project is implement the ACSII art, as well as clean up the code where possible.