Fixed bugs and added difficulties
Found 2 bugs: 1st being that if there are multiple occurences of the same letter in the word, the program will only fill in the first occurence with the user's input. The 2nd bug is that the game didn't end once you won it. Both of these bugs are fixed now. I also added difficulties. For easy, the words are 4 letters long. For normal, 5, and for hard, 6.